analytics_data/requirements.txt
pykafka >= 2.3.0
pymongo >= 4.6.1
python-dotenv == 1.0.0